orchestrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
redis==5.0.8
psycopg2-binary==2.9.9
boto3==1.35.0
pydantic==2.8.2
python-multipart==0.0.9
